requirements.txt
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.29
python-dotenv==1.0.1
psycopg2-binary==2.9.9
scipy==1.12.0
plotly==5.22.0
kaleido==0.2.1
jupyter==1.0.0
seaborn==0.13.2
matplotlib==3.8.4